% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd_photosynthesis.R
\name{calc_assim_rubisco_limited}
\alias{calc_assim_rubisco_limited}
\title{Carboxylation-limited assimilation rate}
\usage{
calc_assim_rubisco_limited(gs, vcmax, par_photosynth)
}
\arguments{
\item{gs}{numeric, Stomatal conductance in mol/m2/s, for e.g. as calculated from \code{\link{calc_gs}}}

\item{vcmax}{numeric, Carboxylation capacity (umol/m2/s)}

\item{par_photosynth}{numeric, A list of photosynthesis parameters, which must include ...... . All concentrations must be converted to partial pressures (Pa)}
}
\value{
Rubisco-limited assimilation rate (umol/m2/s)
}
\description{
Calculates the carboxylation limited CO2 assilimation rate
}
