% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd_photosynthesis.R
\name{calc_assimilation_limiting}
\alias{calc_assimilation_limiting}
\title{Limiting assimilation rate}
\usage{
calc_assimilation_limiting(vcmax, jmax, gs, par_photosynth)
}
\arguments{
\item{vcmax}{numeric, Carboxylation capacity (umol/m2/s)}

\item{jmax}{numeric, Electron transport capacity (umol/m2/s)}

\item{gs}{numeric, Stomatal conductance in mol/m2/s, for e.g. as calculated from \code{\link{calc_gs}}}

\item{par_photosynth}{numeric, A list of photosynthesis parameters, which must include ...... . All concentrations must be converted to partial pressures (Pa)}
}
\value{
Rubisco-limited assimilation rate (umol/m2/s)
}
\description{
Calculates the limiting assimilation rate, the minimum of the 
carboxylation limited and electron-transport-limited rates.
}
