% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_density_h2o.R
\name{calc_density_h2o}
\alias{calc_density_h2o}
\title{Density of water}
\usage{
calc_density_h2o(tc, p)
}
\arguments{
\item{tc}{numeric, air temperature (tc), degrees C}

\item{p}{numeric, atmospheric pressure (p), Pa}
}
\value{
numeric, density of water, kg/m^3
}
\description{
Calculates the density of water as a function of temperature and atmospheric
pressure, using the Tumlirz Equation.
}
\examples{
print("Density of water at 20 degrees C and standard atmospheric pressure:")
print(calc_density_h2o(20, 101325))

}
\references{
F.H. Fisher and O.E Dial, Jr. (1975) Equation of state of
pure water and sea water, Tech. Rept., Marine Physical
Laboratory, San Diego, CA.
}
