% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ftemp_arrh.R
\name{calc_ftemp_arrh}
\alias{calc_ftemp_arrh}
\title{Calculates the Arrhenius-type temperature response}
\usage{
calc_ftemp_arrh(tk, dha, tkref = 298.15)
}
\arguments{
\item{tk}{Temperature (Kelvin)}

\item{dha}{Activation energy (J mol-1)}

\item{tkref}{Reference temperature (Kelvin)}
}
\value{
A numeric value for \eqn{f}
}
\description{
Given a kinetic rate at a reference temperature (argument \code{tkref})
this function calculates its temperature-scaling factor following Arrhenius kinetics.
}
\details{
To correct for effects by temperature following Arrhenius kinetics,
and given a reference temperature \eqn{T0}, \eqn{f} calculates the temperature
scaling. Arrhenius kinetics are described by an equation of form
\eqn{x(T)= exp(c - \Delta Ha / (T R))}. The temperature-correction
function \eqn{f(T, \Delta Ha)} is thus given by \eqn{f=x(T)/x(T0)} which is:
\deqn{
     f = exp( \Delta Ha (T - T0) / (T0 R T_K) )
}
\eqn{\Delta Ha} is given by argument \code{dha}. \eqn{T} is given by argument
\code{tk} and has to be provided in Kelvin. \eqn{R} is the universal gas constant
and is 8.3145 J mol-1 K-1. Note that this is equivalent to 
\deqn{
     f = exp( (\Delta Ha/R) (1/T0 - 1/T) )  
}
}
\examples{
print( paste( "Relative rate change from 25 to 10 degrees Celsius (percent change):", (1.0-calc_ftemp_arrh( 283.15, 100000, tkref = 298.15))*100 ))

}
