% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd_transpiration.R
\name{calc_gs_PM}
\alias{calc_gs_PM}
\title{Stomatal conductance and Transpiration using PM extension}
\usage{
calc_gs_PM(dpsi, psi_soil, par_plant, par_env, PM_params, ...)
}
\arguments{
\item{dpsi}{numeric, soil-to-leaf water potential difference (\eqn{\psi_s-\psi_l}), Mpa}

\item{psi_soil}{numeric, soil water potential, Mpa}

\item{par_plant}{numeric, A list of plant hydraulic parameters, which must include LAI}

\item{par_env}{numeric, A list of environmental parameters which must include wind speed (u), 
friction velocity (ustar), net radiation (nR), viscosity_water, density_water, atmospheric pressure (patm), 
air temperature (tc), vapour pressure deficit (vpd)}
}
\value{
numeric, canopy stomatal conductance in molco2 m-2 s-1
}
\description{
Calculates regulated stomatal conducatnce given the leaf water potential, 
plant hydraulic traits, the environment.
}
