% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kmm.R
\name{calc_kmm}
\alias{calc_kmm}
\title{Michaelis Menten coefficient}
\usage{
calc_kmm(tc, patm)
}
\arguments{
\item{tc}{Temperature, relevant for photosynthesis (deg C)}

\item{patm}{Atmospheric pressure (Pa)}
}
\value{
A numeric value for \eqn{K}
}
\description{
Calculates the Michaelis Menten coefficient of Rubisco-limited assimilation 
as a function of temperature and atmospheric pressure.
}
\details{
The Michaelis-Menten coefficient \eqn{K} of Rubisco-limited 
photosynthesis is determined by the Michalis-Menten constants for 
O2 and CO2 (Farquhar, 1980) according to:
\deqn{
  K = Kc ( 1 + pO_2 / Ko)      
}
where \eqn{Kc} is the Michaelis-Menten constant for CO2 (Pa), \eqn{Ko} is 
the Michaelis-Menten constant for O2 (Pa), and \eqn{pO_2} is the partial 
pressure of oxygen (Pa), calculated as \eqn{0.209476 p}, where \eqn{p} is
given by argument \code{patm}.  \eqn{Kc} and \eqn{Ko} follow a temperature 
dependence, given by the Arrhenius Equation \eqn{f}:
\deqn{
   Kc = Kc25 f(T, \Delta Hkc)
}
\deqn{
   Ko = Ko25 f(T, \Delta Hko)
}
Values \eqn{\Delta Hkc} (79430 J mol-1), \eqn{\Delta Hko} (36380 J mol-1), 
\eqn{Kc25} (39.97 Pa), and \eqn{Ko25} (27480 Pa) are taken from Bernacchi 
et al. (2001) and have been converted from values given therein to units of Pa 
by multiplication with the standard atmosphere (101325 Pa).
}
\examples{
asdf

}
