% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd_numerical_solver.R
\name{calc_vcmax_coordinated_numerical}
\alias{calc_vcmax_coordinated_numerical}
\title{Coordinated Vcmax}
\usage{
calc_vcmax_coordinated_numerical(aj, ci, par_photosynth)
}
\arguments{
\item{aj}{numeric, electron-transport limited assimilation rate (umol/m2/s)}

\item{ci}{numeric, leaf-internal CO2 concentration, converted to partial pressure (Pa)}

\item{par_photosynth}{numeric, A list of photosynthesis parameters, which must include ...... . All concentrations must be converted to partial pressures (Pa)}
}
\value{
Coordinated Vcmax (umol/m2/s)
}
\description{
Calculates the carboxylation capacity, as coordinated to a given electron-transport limited assimilation rate.
}
