% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd_analytical_solver_3.R
\name{chi_jmax_limited}
\alias{chi_jmax_limited}
\title{Electron-transport limited \eqn{\chi}}
\usage{
chi_jmax_limited(par_photosynth, par_cost)
}
\arguments{
\item{par_photosynth}{numeric, A list of photosynthesis parameters, which must include ...... . All concentrations must be converted to partial pressures (Pa)}

\item{par_cost}{numeric, A list of cost parameters}
}
\value{
\eqn{\chi}
}
\description{
Analytically calculate \eqn{\chi} in the case of strong Jmax limitation
}
