% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subroutines.R
\name{co2_to_ca}
\alias{co2_to_ca}
\title{CO2 partial pressure}
\usage{
co2_to_ca(co2, patm)
}
\arguments{
\item{co2}{Atmospheric CO2 concentration (ppm)}

\item{patm}{Atmospheric pressure (Pa).}
}
\value{
CO2 partial pressure in Pa.
}
\description{
Calculates CO2 partial pressure from concentration in ppm.
}
