% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd_numerical_solver.R
\name{fn_profit}
\alias{fn_profit}
\title{Profit Function}
\usage{
fn_profit(
  par,
  psi_soil,
  par_cost,
  par_photosynth,
  par_plant,
  par_env,
  do_optim = FALSE,
  opt_hypothesis,
  gs_approximation
)
}
\arguments{
\item{par}{numeric, A vector of variables to be optimized, namely, \eqn{c(log(Jmax), \Delta\psi)}}

\item{psi_soil}{numeric, soil water potential (Mpa)}

\item{par_cost}{numeric, List of named cost parameters: list(alpha = xx, gamma = xx)}

\item{par_photosynth}{numeric, Photosynthesis parameters}

\item{par_env}{numeric, Environmental parameters}

\item{opt_hypothesis}{character, Either "Lc" or "PM"}

\item{gs_approximation}{character, Either "Ohm" or "PM". Ohm is used when gs is calculated as ohm approximation and PM is used when gs is calculated using Penman-Monteith}
}
\value{
Net assimilation rate after accounting for costs (profit) (umol/m2/s)
}
\description{
The profit function passed to the optimizer. It calculates 
the profit, defined as \eqn{(A - \alpha Jmax - \gamma\Delta\psi^2)}
}
