% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd_instantaneous_solver.R
\name{fn_profit_instantaneous}
\alias{fn_profit_instantaneous}
\title{Profit Function}
\usage{
fn_profit_instantaneous(
  par,
  jmax,
  vcmax,
  psi_soil,
  par_cost,
  par_photosynth,
  par_plant,
  par_env,
  gs_approximation = gs_approximation,
  do_optim = F
)
}
\arguments{
\item{par}{numeric, A vector of variables to be optimized, namely, \eqn{c(log(Jmax), \Delta\psi)}}

\item{jmax}{numeric, Jmax (potentially acclimated to previous conditions) (umol/m2/s)}

\item{vcmax}{numeric, Vcmax (potentially coordinated to previously acclimated Jmax) (umol/m2/s)}

\item{psi_soil}{numeric, soil water potential (Mpa)}

\item{par_cost}{numeric, List of named cost parameters: list(alpha = xx, gamma = xx)}

\item{par_photosynth}{numeric, Photosynthesis parameters}

\item{par_env}{numeric, Environmental parameters}

\item{opt_hypothesis}{character, Either "Lc" or "PM"}
}
\value{
Net assimilation rate after accounting for costs (profit) (umol/m2/s)
}
\description{
The profit function passed to the \eqn{\Delta\psi} optimizer. It calculates 
the profit, defined as \eqn{(A - \gamma\Delta\psi^2)}
}
