% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subroutines.R
\name{optimal_chi}
\alias{optimal_chi}
\title{Calculate optimal Chi}
\usage{
optimal_chi(kmm, gammastar, ns_star, ca, vpd, beta)
}
\arguments{
\item{kmm}{Michaelis Menten coefficient}

\item{gammastar}{CO2 compensation point}

\item{ns_star}{viscosity correction factor for water}

\item{ca}{ambient CO2}

\item{vpd}{vapour pressure deficit}

\item{beta}{beta parameter}
}
\value{
optimal chi value
}
\description{
Optimal chi routine
}
