% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyd_vulnerability_curve.R
\name{P}
\alias{P}
\alias{Pprime}
\alias{Pprimeprime}
\title{Hydraulic vulnerability curve}
\usage{
P(psi, psi50, b)

Pprime(psi, psi50, b)

Pprimeprime(psi, psi50, b)
}
\arguments{
\item{psi}{numeric, Water potential (Mpa)}

\item{psi50}{numeric, Water potential at which 50\% conductivity is lost (Mpa)}

\item{b}{numeric, Slope of the vulnerability curve}
}
\value{
Fractional conductivity
}
\description{
Calculates the conductivity of the hydraulic pathway (stem or leaf) as a function 
of water potential, using a 2-parameter Weibull function to describe the hydraulic 
vulnerability curve:
\deqn{P(\psi) = (1/2)^((\psi/\psi_50)^b) }
}
\section{Functions}{
\itemize{
\item \code{Pprime}: First derivative of the vulnerability curve

\item \code{Pprimeprime}: Second derivative of the vulnerability curve
}}

