% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpmodel_hydraulics_analytical.R
\name{pmodel_hydraulics_analytical}
\alias{pmodel_hydraulics_analytical}
\title{The hydraulic p-model (HyPE), analytical version}
\usage{
pmodel_hydraulics_analytical(
  tc,
  ppfd,
  vpd,
  co2,
  elv,
  fapar,
  kphio,
  psi_soil,
  rdark = 0,
  par_plant,
  par_cost
)
}
\description{
Calculates the carboxylation capacity, as coordinated to a given electron-transport limited assimilation rate.
}
