% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpmodel_hydraulics_instantaneous.R
\name{pmodel_hydraulics_instantaneous}
\alias{pmodel_hydraulics_instantaneous}
\title{The hydraulic p-model (HyPE), numerical version}
\usage{
pmodel_hydraulics_instantaneous(
  tc,
  ppfd,
  vpd,
  co2,
  elv,
  fapar,
  kphio,
  psi_soil,
  rdark = 0,
  par_plant,
  par_cost = NULL,
  vcmax,
  jmax,
  opt_hypothesis = "PM"
)
}
\description{
Instantaneous photosynthesis rates for a given Vcmax and Jmax.
}
