% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpmodel_hydraulics_numerical.R
\name{pmodel_hydraulics_numerical}
\alias{pmodel_hydraulics_numerical}
\title{The hydraulic p-model (HyPE), numerical version}
\usage{
pmodel_hydraulics_numerical(
  tc,
  ppfd,
  vpd,
  co2,
  elv,
  fapar,
  kphio,
  psi_soil,
  rdark = 0,
  par_plant,
  par_cost = NULL,
  opt_hypothesis = "PM"
)
}
\description{
Calculates the carboxylation capacity, as coordinated to a given electron-transport limited assimilation rate.
}
