% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst_rpmodel.R
\name{QUADM}
\alias{QUADM}
\title{Minor quadratic root}
\usage{
QUADM(A, B, C)
}
\arguments{
\item{A}{a parameter in the quadratic equation}

\item{B}{b parameter in the quadratic equation}

\item{C}{c parameter in the quadratic equation}
}
\value{
minor quardratic root
}
\description{
Solves quadratic equation given by y = a*x^2 + bx + c
}
