% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUADP.R
\name{QUADP}
\alias{QUADP}
\title{Quadratic Solver}
\usage{
QUADP(A, B, C)
}
\arguments{
\item{A}{numeric, Stomatal conductance in mol/m2/s, for e.g. as calculated from \code{\link{calc_gs}}}

\item{B}{numeric, Carboxylation capacity (umol/m2/s)}

\item{C}{numeric, A list of photosynthesis parameters, which must include ...... . All concentrations must be converted to partial pressures (Pa)}
}
\value{
root
}
\description{
Quadratic Solver
}
