% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpmodel.R
\name{rpmodel}
\alias{rpmodel}
\title{P-model}
\usage{
rpmodel(tc, vpd, co2, elv, kphio, beta = 146, fapar = NA, ppfd = NA,
  c4 = FALSE, method_optci = "prentice14", method_jmaxlim = "wang17",
  do_ftemp_kphio = TRUE, returnvar = NULL)
}
\arguments{
\item{tc}{Temperature, relevant for photosynthesis (deg C)}

\item{vpd}{Vapour pressure deficit (Pa)}

\item{co2}{Atmospheric CO2 concentration (ppm)}

\item{elv}{Elevation above sea-level (m.a.s.l.)}

\item{kphio}{Quantum yield efficiency parameter}

\item{beta}{Unit cost ratio. Defaults to 146.0.}

\item{fapar}{(Optional) Fraction of absorbed photosynthetically active radiation (unitless, defaults to \code{NA})}

\item{ppfd}{(Optional) Photosynthetic photon flux density (mol/m2, defaults to \code{NA})}

\item{c4}{(Optional) A logical value specifying whether the C3 or C4 photosynthetic pathway is followed. Defaults to \code{method_optci="c4=FALSE"}. 
If \code{method_optci="c4=TRUE"}, ci is assumed to be very large and \code{lue = kphio * fapar * ppfd}.}

\item{method_optci}{(Optional) A character string specifying which method is to be used for calculating optimal ci:ca. Defaults to \code{"prentice14"}. 
Available also \code{"prentice14_num"} for a numerical solution to the same optimization criterium as used for \code{"prentice14"}.}

\item{method_jmaxlim}{(Optional) A character string specifying which method is to be used for factoring in Jmax limitation. Defaults to \code{"wang17"}, 
based on Wang Han et al. 2017 Nature Plants and (Smith 1937). Available is also \code{"smith19"}, following the method by Smith et al., 2019 Ecology Letters, 
and \code{"none"} for ignoring effects of Jmax limitation.}

\item{do_ftemp_kphio}{(Optional) A logical specifying whether temperature-dependence of quantum yield efficiency after Bernacchi et al., 2003 PCE 
is to be accounted for. Defaults to \code{TRUE}.}

\item{returnvar}{(Optional) A character string of vector of character strings specifying which variables are to be returned (see return below).}
}
\value{
A named list of numeric values with 
\itemize{
        \item \code{gammastar}: photorespiratory compensation point, (Pa). It is calculated as:
        \item \code{kmm}: Michaelis-Menten coefficient for photosynthesis (Pa)
        \item \code{ci}: leaf-internal partial pressure, (Pa)
        \item \code{chi}: = ci/ca, leaf-internal to ambient CO2 partial pressure, ci/ca (unitless)
        \item \code{iwue}: intrinsic water use efficiency (unitless)
        \item \code{lue}: light use efficiency (mol CO2 / mol photon)
        \item \code{gpp}: gross primary productivity (g C m-2). Calculated only if \code{fapar} and \code{ppfd} are not \code{NA}.
        \item \code{vcmax}: maximum carboxylation capacity per unit ground area (mol CO2 m-2 s-1)
        \item \code{vcmax25}: Vcmax25 (Vcmax normalized to 25 deg C) (mol CO2 m-2 s-1)
        \item \code{vcmax_unitfapar}: Vcmax per fAPAR (mol CO2 m-2 s-1)
        \item \code{vcmax_unitiabs}: Vcmax per unit absorbed light (xxx units)
        \item \code{rd}: Dark respiration (mol CO2 m-2 s-1)
        \item \code{rd_unitfapar}: Dark respiration per fAPAR (mol CO2 m-2 s-1)
        \item \code{rd_unitiabs}: Dark respiration per unit absorbed light (mol CO2 m-2 s-1)
        \item \code{actnv}: Active metabolic leaf N (canopy-level), mol N/m2-ground
        \item \code{actnv_unitfapar}: Active metabolic leaf N (leaf-level, top of canopy), mol N/m2-leaf
        \item \code{actnv_unitiabs}: Active metabolic leaf N per unit absorbed light, mol N/m2/mol
}
}
\description{
R implementation of the P-model and its corrolary predictions (Prentice et al., 2014; Han et al., 2017)
}
\examples{
out_rpmodel <- rpmodel( tc=10, vpd=300, co2=300, elv=300, kphio=0.06 )

}
