% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subroutines.R
\name{QUADP}
\alias{QUADP}
\title{Larger quadratic root}
\usage{
QUADP(A, B, C)
}
\arguments{
\item{A}{a parameter in the quadratic equation}

\item{B}{b parameter in the quadratic equation}

\item{C}{c parameter in the quadratic equation}
}
\value{
larger quardratic root
}
\description{
Solves quadratic equation given by y = a*x^2 + bx + c
}
